updating code according to DAO changes. no working tests yet.

+++ whois-query/src/main/java/net/ripe/db/whois/query/query/Query.java
@@ -241,7 +241,7 @@ public class Query {
         return queryParser.hasOption(QueryFlag.SHOW_INTERNAL_VERSION);
     }
 
-    public long getObjectInternalVersion() {
+    public int getObjectInternalVersion() {
         return getObjectVersion(QueryFlag.SHOW_INTERNAL_VERSION);
     }
